rename arbiter handle module
