+++ crates/chain/src/lib.rs
@@ -1,5 +1,5 @@
 pub mod chain;
-pub(crate) mod clonable_join_handle;
+pub(crate) mod arbiter_handle;
 pub mod vdf;
 
 pub use chain::*;
